+++ packages/mobile/e2e/utils/press.js
@@ -1,14 +1,5 @@
-import info from './info'
-
-const { ios } = info
-
-const press = async (element, double = false) => {
-  if (ios) {
-    await element.tap()
-  } else {
-    if (double) await element.longPress() // Idle
-    await element.longPress()
-  }
+const press = async (element) => {
+  await element.tap()
 }
 
 export default press