+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+[unreleased]
 
-[Unreleased]
+* Fixed mobile bugs - joining by QR code and not showing username taken screen for user who has unique name
+
+* Shorter dots-placeholder for invite link
+
+* Removed registration attempts selector and corresponding usage.
 
 * Revert adjusting bootstrap scripts for developing on Windows